backend/requirements.txt
# backend/requirements.txt
fastapi==0.115.2
uvicorn==0.32.0
mysql-connector-python==9.1.0
bcrypt==4.2.0
pyjwt==2.9.0
httpx==0.27.2
pandas==2.2.3
botocore==1.34.69
aiobotocore==2.12.3
requests==2.32.3
idna==3.7
python-dateutil==2.9.0
